+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [] -
 ### Changed
 - Speed-up CI tests by caching installation of libs and splitting tests into randomized groups using pytest-test-groups
-- Mock Ensembl services (gene conversion and liftover) used in patient parsing tests
+- Mock Ensembl services (gene conversions and liftover) used all tests
 - Mock the download of files containing phenotype resources in tests
 
 ## [3.0] - 2021-12-27